refactor(Entire): change Umpire's location in package
change the location of Umpire in package because it makes more sense when it is in model package
(Umpire가 model package에 있는 것이 맞다고 생각 되기 때문에 Umpire를 model 패키지로 옯긴다)
